product_details/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2.13
django-cors-headers==4.3.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==3.4
PyJWT==1.7.1
python-decouple==3.3
pytz==2023.3.post1
requests==2.27.1
sqlparse==0.4.4
urllib3==1.26.18
